refactor: convert all blocks to typescript
