+++ packages/form-builder/src/App.tsx
@@ -2,17 +2,17 @@ import {ShortcutProvider} from '@wordpress/keyboard-shortcuts';
 import BlockEditorContainer from './containers/BlockEditorContainer';
 import {FormStateProvider} from './stores/form-state';
 import {Storage} from './common';
-import type {Block} from '@givewp/form-builder/types';
 
 import './App.scss';
 
 import defaultBlocks from './blocks.json';
 import Feedback from '@givewp/form-builder/feedback';
+import {BlockInstance} from '@wordpress/blocks';
 
 const {blocks: initialBlocks, formSettings: initialFormSettings} = Storage.load();
 
 const initialState = {
-    blocks: initialBlocks || (defaultBlocks as Block[]),
+    blocks: initialBlocks || (defaultBlocks as BlockInstance[]),
     settings: {
         ...initialFormSettings,
     },
