+++ packages/form-builder/src/blocks/fields/amount/inspector/delete-button.tsx
@@ -1,4 +1,4 @@
-import {__} from "@wordpress/i18n";
+import {__} from '@wordpress/i18n';
 
 const DeleteButton = (props) => {
     return (
