+++ packages/form-builder/src/blocks/fields/amount/inspector/index.tsx
