+++ packages/form-builder/src/blocks/fields/index.ts
@@ -5,12 +5,13 @@ import email from './email';
 import paymentGateways from './payment-gateways';
 import donationSummary from './donation-summary';
 import amount from './amount';
+import {FieldBlock} from '@givewp/form-builder/types';
 
 /**
  * @note Blocks in the appender are listed in the order that the blocks are registered.
  */
 
-const fieldBlocks = [
+const fieldBlocks: FieldBlock[] = [
     field,
     company,
     donorName,
