+++ packages/form-builder/src/blocks/fields/settings/DefaultFieldSettings.jsx
@@ -17,6 +17,8 @@ export default function DefaultFieldSettings({attributes, setAttributes, onLabel
                         required={isRequired}
                         className={requiredClass}
                         readOnly
+                        onChange={null}
+                        value={placeholder}
                     />
                 ) : (
                     <select>
