+++ packages/form-builder/src/blocks/fields/settings/index.tsx
@@ -1,14 +1,14 @@
 import {__} from '@wordpress/i18n';
 import DefaultFieldSettings from './DefaultFieldSettings';
+import {BlockConfiguration} from '@wordpress/blocks';
 
-const defaultSettings = {
+const defaultSettings: BlockConfiguration = {
     title: __('Field', 'custom-block-editor'),
-
+    category: 'text',
     supports: {
         html: false, // Removes support for an HTML mode.
         multiple: true,
     },
-
     attributes: {
         fieldName: {
             type: 'string',
