+++ packages/form-builder/src/common/storage/interface.ts
@@ -1,6 +1,5 @@
 import {FormData} from '../../types/formData';
-import {Block} from '../../types/block';
-import {FormSettings} from "@givewp/form-builder/types";
+import {FormSettings} from '@givewp/form-builder/types';
 
 export interface StorageDriver {
     save({blocks, formSettings}: FormData): Promise<FormSettings>;
