+++ packages/form-builder/src/components/sidebar/secondary.tsx
@@ -1,10 +1,10 @@
 import {__} from '@wordpress/i18n';
-import FieldTypesList from "./panels/field-types-list";
-import BlockListTree from "./panels/block-list-tree";
+import FieldTypesList from './panels/FieldTypesList';
+import BlockListTree from './panels/block-list-tree';
 
 type PropTypes = {
-    selected: string
-}
+    selected: string;
+};
 
 const Sidebar = ({selected}: PropTypes) => {
