+++ packages/form-builder/src/index.tsx
@@ -1,18 +1,17 @@
 import React from 'react';
 import {createRoot} from 'react-dom/client';
-import {BlockConfiguration, registerBlockType} from '@wordpress/blocks';
+import {registerBlockType} from '@wordpress/blocks';
 
 import App from './App';
 
 import sectionBlocks, {sectionBlockNames} from './blocks/section';
 import fieldBlocks from './blocks/fields';
 import elementBlocks from './blocks/elements';
+import {FieldBlock} from '@givewp/form-builder/types';
 
-sectionBlocks.map(({name, settings}: { name: string; settings: BlockConfiguration }) =>
-    registerBlockType(name, settings)
-);
+sectionBlocks.map(({name, settings}: FieldBlock) => registerBlockType(name, settings));
 
-[...fieldBlocks, ...elementBlocks].map(({name, settings}: { name: string; settings: BlockConfiguration }) =>
+[...fieldBlocks, ...elementBlocks].map(({name, settings}: FieldBlock) =>
     registerBlockType(name, {...settings, parent: sectionBlockNames})
 );
