+++ packages/form-builder/src/types/index.ts
@@ -1,4 +1,4 @@
-export type {Block} from './block';
+export type {FieldBlock} from './block';
 export type {FormData} from './formData';
 export type {FormDesign} from './formDesign';
 export type {FormSettings} from './formSettings';
